pyproject.toml
[project]
name = "docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lancedb>=0.29.2",
    "lance-namespace>=0.5.2",
    "pandas>=2.3.3",
    "polars>=1.35.2",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "Pillow>=11.0.0",
]
